Initial test restructuring
- Removed obsolete `Command` and `Help` tests
- Re-arranged `Module` tests and updated namespaces and class names
