--- test/ScanForErrorMiddleware/HelpTest.php
@@ -1,62 +1,0 @@
-<?php
-/**
- * @see       https://github.com/zendframework/zend-expressive-tooling for the canonical source repository
- * @copyright Copyright (c) 2016-2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-tooling/blob/master/LICENSE.md New BSD License
- */
-
-namespace ZendTest\Tooling\ScanForErrorMiddleware;
-
-use PHPUnit\Framework\TestCase;
-use Prophecy\Argument;
-use Zend\Expressive\Tooling\ScanForErrorMiddleware\Help;
-use Zend\Stdlib\ConsoleHelper;
-
-class HelpTest extends TestCase
-{
-    public function testWritesHelpMessageToConsoleUsingCommandProvidedAtInstantiationAndResourceAtInvocation()
-    {
-        $resource = fopen('php://temp', 'wb+');
-
-        $console = $this->prophesize(ConsoleHelper::class);
-        $console
-            ->writeLine(
-                Argument::that(function ($message) {
-                    return false !== strpos($message, 'scan-for-error-middleware');
-                }),
-                true,
-                $resource
-            )
-            ->shouldBeCalled();
-
-        $command = new Help(
-            'scan-for-error-middleware',
-            $console->reveal()
-        );
-
-        $this->assertNull($command($resource));
-    }
-
-    public function testTruncatesCommandToBasenameIfItIsARealpath()
-    {
-        $resource = fopen('php://temp', 'wb+');
-
-        $console = $this->prophesize(ConsoleHelper::class);
-        $console
-            ->writeLine(
-                Argument::that(function ($message) {
-                    return false !== strpos($message, basename(__FILE__));
-                }),
-                true,
-                $resource
-            )
-            ->shouldBeCalled();
-
-        $command = new Help(
-            realpath(__FILE__),
-            $console->reveal()
-        );
-
-        $this->assertNull($command($resource));
-    }
-}
